requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
pypdf>=4.1.0
openai>=1.14.0
python-dotenv>=1.0.1
httpx>=0.27.0
tiktoken>=0.6.0
pydantic>=2.6.4
pydantic-settings>=2.2.1
qdrant-client>=1.8.0
prometheus-fastapi-instrumentator>=7.0.0
